Refactor vital data post processing.
